# swagger.yaml
openapi: 3.0.0
info:
  title: Foot City
  contact: {}
  version: '1.0'
servers:
- url: http://example.com
  variables: {}
paths:
...
  /review:
    get:
      tags:
      - Review
      summary: Get All Reviews
      description: This request gets all reviews available in database
      operationId: GetAllReviews
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  reviews:
                  - _id: 63362f5be704ab2653e5eda2
                    title: Good Rating
                    comment: This is a really good product
                    rating: 0
                    user:
                      _id: 6329108dddd6b7fc854ccd02
                      firstName: elkanah
                    product: 63362bc9e704ab2653e5ed9c
                    replies: []
                    createdAt: 2022-09-29T23:50:51.425Z
                    updatedAt: 2022-09-29T23:50:51.425Z
                    __v: 0
                  count: 1
              example:
                reviews:
                - _id: 63362f5be704ab2653e5eda2
                  title: Good Rating
                  comment: This is a really good product
                  rating: 0
                  user:
                    _id: 6329108dddd6b7fc854ccd02
                    firstName: elkanah
                  product: 63362bc9e704ab2653e5ed9c
                  replies: []
                  createdAt: 2022-09-29T23:50:51.425Z
                  updatedAt: 2022-09-29T23:50:51.425Z
                  __v: 0
                count: 1
      deprecated: false
      security: []
    post:
      tags:
      - Review
      summary: Create Review
      description: This request creates a review and associates it to a particular product that exists. A user cannot have more than one review for a particular product.
      operationId: CreateReview
      parameters: []
      requestBody:
        description: The product field value must be the id of a product that exist
        content:
          text/plain:
            schema:
              type: object
              description: The product field value must be the id of a product that exist
              example:
                title: Good Rating
                comment: This is a really good product
                product: 63362bc9e704ab2653e5ed9c
                rating: 4
            example:
              title: Good Rating
              comment: This is a really good product
              product: 63362bc9e704ab2653e5ed9c
              rating: 4
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  review:
                    title: Good Rating
                    comment: This is a really good product
                    rating: 4
                    user: 6329108dddd6b7fc854ccd02
                    product: 63362bc9e704ab2653e5ed9c
                    _id: 633632b0e704ab2653e5edb2
                    replies: []
                    createdAt: 2022-09-30T00:05:04.517Z
                    updatedAt: 2022-09-30T00:05:04.517Z
                    __v: 0
              example:
                review:
                  title: Good Rating
                  comment: This is a really good product
                  rating: 4
                  user: 6329108dddd6b7fc854ccd02
                  product: 63362bc9e704ab2653e5ed9c
                  _id: 633632b0e704ab2653e5edb2
                  replies: []
                  createdAt: 2022-09-30T00:05:04.517Z
                  updatedAt: 2022-09-30T00:05:04.517Z
                  __v: 0
      deprecated: false
  /review/{review_id}:
    get:
      tags:
      - Review
      summary: Get Single Review
      description: This request gets a single review that matches the id of an existing review
      operationId: GetSingleReview
      parameters:
      - name: review_id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  review:
                    _id: 633632b0e704ab2653e5edb2
                    title: Good Rating
                    comment: This is a really good product
                    rating: 4
                    user:
                      _id: 6329108dddd6b7fc854ccd02
                      firstName: elkanah
                      email: maxetoo@gmail.com
                    product: 63362bc9e704ab2653e5ed9c
                    replies: []
                    createdAt: 2022-09-30T00:05:04.517Z
                    updatedAt: 2022-09-30T00:05:04.517Z
                    __v: 0
              example:
                review:
                  _id: 633632b0e704ab2653e5edb2
                  title: Good Rating
                  comment: This is a really good product
                  rating: 4
                  user:
                    _id: 6329108dddd6b7fc854ccd02
                    firstName: elkanah
                    email: maxetoo@gmail.com
                  product: 63362bc9e704ab2653e5ed9c
                  replies: []
                  createdAt: 2022-09-30T00:05:04.517Z
                  updatedAt: 2022-09-30T00:05:04.517Z
                  __v: 0
      deprecated: false
    patch:
      tags:
      - Review
      summary: Update Review
      description: This request updates a particular existing review based on it's id
      operationId: UpdateReview
      parameters:
      - name: review_id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: ''
        content:
          text/plain:
            schema:
              type: object
              example:
                title: Great product
                comment: Just really top notch
                rating: 5
            example:
              title: Great product
              comment: Just really top notch
              rating: 5
        required: true
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  review:
                    _id: 633632b0e704ab2653e5edb2
                    title: Great product
                    comment: Just really top notch
                    rating: 5
                    user: 6329108dddd6b7fc854ccd02
                    product: 63362bc9e704ab2653e5ed9c
                    replies: []
                    createdAt: 2022-09-30T00:05:04.517Z
                    updatedAt: 2022-09-30T00:09:42.394Z
                    __v: 0
              example:
                review:
                  _id: 633632b0e704ab2653e5edb2
                  title: Great product
                  comment: Just really top notch
                  rating: 5
                  user: 6329108dddd6b7fc854ccd02
                  product: 63362bc9e704ab2653e5ed9c
                  replies: []
                  createdAt: 2022-09-30T00:05:04.517Z
                  updatedAt: 2022-09-30T00:09:42.394Z
                  __v: 0
      deprecated: false
    delete:
      tags:
      - Review
      summary: Delete Review
      description: This request deletes a particular review for a specific product
      operationId: DeleteReview
      parameters:
      - name: review_id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  msg: Review deleted successfully!
              example:
                msg: Review deleted successfully!
      deprecated: false
...
  /review/productReviews/{product_id}:
    get:
      tags:
      - Review
      summary: Get Product Reviews
      description: This request gets all the reviews associated to a particular product
      operationId: GetProductReviews
      parameters:
      - name: product_id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
          content:
            text/plain:
              schema:
                type: object
                example:
                  reviews:
                  - _id: 633632b0e704ab2653e5edb2
                    title: Great product
                    comment: Just really top notch
                    rating: 5
                    user: 6329108dddd6b7fc854ccd02
                    product: 63362bc9e704ab2653e5ed9c
                    replies: []
                    createdAt: 2022-09-30T00:05:04.517Z
                    updatedAt: 2022-09-30T00:18:25.376Z
                    __v: 0
                  count: 1
              example:
                reviews:
                - _id: 633632b0e704ab2653e5edb2
                  title: Great product
                  comment: Just really top notch
                  rating: 5
                  user: 6329108dddd6b7fc854ccd02
                  product: 63362bc9e704ab2653e5ed9c
                  replies: []
                  createdAt: 2022-09-30T00:05:04.517Z
                  updatedAt: 2022-09-30T00:18:25.376Z
                  __v: 0
                count: 1
      deprecated: false
...
components:
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...